CRUD - Agregado con estilos

+++ src/app/components/list-product/list-product.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Producto } from 'src/app/interfaces/producto';
 import { ProductoService } from 'src/app/service/producto.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-product',
@@ -20,7 +21,25 @@ export class ListProductComponent {
 
   }
 
-onDeleteProduct(producto:Producto){
-  this.productoService.deleteProduct(producto)
-}
+  onDeleteProduct(producto: Producto) {
+
+    Swal.fire({
+      title: 'Desea eliminar producto',
+      icon: 'error',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Si, Eliminar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Exitoso!',
+          'Has Eliminado producto exitosamente',
+          'success'
+        )
+        this.productoService.deleteProduct(producto)
+      }
+    })
+
+  }
 }
